chore: remove recreateTables from runtime
